\setmathfont: Add \minus and · (U+00B7), remove *
Related issue for \minus: https://github.com/wspr/unicode-math/issues/451

Adding · (U+00B7 MIDDLE DOT) because it is sometimes used as an alternative
for ⋅ (U+22C5 DOT OPERATOR).

Removing * (asterisk) because it is not used in math and the symbol that’s
actually needed (∗, U+2217 ASTERISK OPERATOR) is not present in Meta
Office Pro.